extension: Move interfaces to separated files
